refactor WebViewActivity
WebViewActivity is primarily used to display post and page previews, as
well as display the WordPress admin dashboard.  It looks like it
previously was also used for the WordPress.com reader?

This change refactors WebViewActivity into several classes to provide a
cleaner API, making it easier to add additional WebViews (coming soon),
and removes a significant chunk of duplicated and/or obsolete code.
There *should* be no user-visible changes here.

This change also starts putting Activities into the 'ui' package, as
things are getting a little crowded in org.wordpress.android.