Catch null loc values when getting post from DB
We were just grabbing the lat/lng values set in the db and setting as
is. For posts without a location, this meant that we would set their
lat/lng to 0,0 which is a valid location.

We now check for null db column entries before setting the location
value for the post to prevent that.