Call getLocation only for posts without location
Post.isNew has a bug currently, which causes it to return
true for almost all posts (see #1463). Instead of prompting the
auto-location detection for isNew posts, so it only for posts that don't
have a location so we don't end up unnecessarily overriding things.

We will override this behavior soon to make the location setting
on-demand.