Geotagging UI Refresh
Make the geotagging for a post on-demand instead of automatically
adding location to any new post or post without a location.

When a post doesn't have a location, users have to explicitly choose to
add a location. They can use their current location, if available, or
search for a location.

The experience now more closely mirrors the updated "add location"
experience on WordPress.com.

see #870