Wait for nested AsyncTask before ending the test
So, there's a nested AsyncTask in RefreshBlogContentTask that typically executes after RefreshBlogContentTask itself finishes. But, the test tears down at that time causing a race condition, ending up with the nested GetPostFormatsTask picking up the cleared Factories instead of the test ones.

As a fix, the tests now use a trick to wait for the nested AsyncTask to finish: post a new Runnable to the SERIAL_EXECUTOR that handles the AsyncTasks in question. The Runnable triggers the CountDownLatch thus letting the tests finish gracefully.