Repaint the prefs list dialog on select Fixes #3958
The adapter can (and that's what it does) call getView() with null convertview more times than the list items count so, it's not reliable to assume that we know which RadioButton was last selected. Instead, notify that the adapter's data has changed (it has, the "selected list item" index is part of the data) and let the list repaint.