Better handling of rotation
By the time the two async operations complete (Gravatar upload, reloading the image) the Fragment can be a new instance if rotation has happened in the meantime. So, have the callbacks send messages (EventBus) so the new Fragments can catch them.

While at it, keep the progress bar state on rotation.